AA: Extract GetToken logic as a crate
