+++ attestation-agent/token/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "token"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+kbs_protocol = { path = "../kbs_protocol", optional = true }
+tokio = { workspace = true, features = ["fs"] }
+toml.workspace = true
+
+[features]
+default = ["kbs"]
+kbs = ["kbs_protocol/background_check"]
